[a11y] Moved the desktop accessibility-related files into an `a11y` sub-package.
