+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/platform/a11y/ComposeSceneAccessible.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.platform
+package androidx.compose.ui.platform.a11y
 
-import androidx.compose.ui.platform.ComposeSceneAccessible.ComposeSceneAccessibleContext
+import androidx.compose.ui.platform.a11y.ComposeSceneAccessible.ComposeSceneAccessibleContext
 import java.awt.*
 import java.awt.event.FocusListener
 import java.util.*
